Swap grid for Field in _search_time_index
This is done since the Field object holds either an xr.DataArray or
ux.UxDataArray which contain the grid, dimensions, and coordinates. With
this, the time argument is now set to be a datetime object.
